Make cycleCounter message level so it works in async context.
Also, check the owner in the WHERE clause of the requeue query. You must
own the message to be able to requeue it.

Adding milliseconds to heartbeatDate and remove isProcessing index

With isProcessing indexed we were getting deadlocks. It appears the
index would be scanned and while that happened another query would grab
a lock on a record before the one scanning could. I am still a bit
confused about it tbh.

https://www.xaprb.com/blog/2006/08/03/a-little-known-way-to-cause-a-database-deadlock/

Dealing with deadlocks on restore and polling.

With our current setup deadlocks are always  possible so we have to deal
with them. I added logic to continue on when deadlocks occur during
polling and restoring. I also added a deadlockCountLimit which if
consecutive deadlocks occur it will emit an error once it reaches that
count. Also, update the tests.